wake: `foo` is now a literal regular expression
These get checked for correctness at lexing time.
